add maintenance status note to readme
@kitten @sofiapoh We're adding maintenance status to all our OSS. I think this project should be listed as "Active" because of its association with `react-live`, but it might also be listed as "Stable" if you think that's more appropriate. If so, please replace the note I added with:

```
Maintenance Status: Stable

Formidable is not planning to develop any new features for this project. We are still responding to bug reports and security concerns. We are still welcoming PRs for this project, but PRs that include new features should be small and easy to integrate and should not include breaking changes.
```